+++ packages/integration-sdk-runtime/src/storage/FileSystemGraphObjectStore/FileSystemGraphObjectStore.ts
@@ -8,11 +8,11 @@ import {
   GraphObjectIteratee,
   IntegrationDuplicateKeyError,
   IntegrationMissingKeyError,
+  GraphObjectLookupKey,
 } from '@jupiterone/integration-sdk-core';
 
 import { flushDataToDisk } from './flushDataToDisk';
 import { BucketMap } from './BucketMap';
-import { DuplicateKeyTracker } from '../../execution/jobState';
 import {
   iterateEntityTypeIndex,
   iterateRelationshipTypeIndex,
@@ -24,31 +24,18 @@ export const GRAPH_OBJECT_BUFFER_THRESHOLD = 500; // arbitrarily selected, subje
 // to ensure that only one operation can be performed at a time.
 const BINARY_SEMAPHORE_CONCURRENCY = 1;
 
-export interface FileSystemGraphObjectStoreParams {
-  duplicateKeyTracker: DuplicateKeyTracker;
-}
-
 export class FileSystemGraphObjectStore {
-  private duplicateKeyTracker: DuplicateKeyTracker;
-
   semaphore: Sema;
   entityStorageMap: BucketMap<Entity>;
   relationshipStorageMap: BucketMap<Relationship>;
 
-  constructor({ duplicateKeyTracker }: FileSystemGraphObjectStoreParams) {
+  constructor() {
     this.entityStorageMap = new BucketMap();
     this.relationshipStorageMap = new BucketMap();
-    this.duplicateKeyTracker = duplicateKeyTracker;
     this.semaphore = new Sema(BINARY_SEMAPHORE_CONCURRENCY);
   }
 
   async addEntities(storageDirectoryPath: string, newEntities: Entity[]) {
-    newEntities.forEach((e) => {
-      this.duplicateKeyTracker.registerKey(e._key, {
-        _type: e._type,
-      });
-    });
-
     this.entityStorageMap.add(storageDirectoryPath, newEntities);
 
     if (this.entityStorageMap.totalItemCount >= GRAPH_OBJECT_BUFFER_THRESHOLD) {
@@ -70,19 +57,8 @@ export class FileSystemGraphObjectStore {
     }
   }
 
-  async getEntity(_key: string): Promise<Entity> {
+  async getEntity({ _key, _type }: GraphObjectLookupKey): Promise<Entity> {
     await this.flushEntitiesToDisk();
-    const graphObjectMetadata = this.duplicateKeyTracker.getGraphObjectMetadata(
-      _key,
-    );
-
-    if (!graphObjectMetadata) {
-      throw new IntegrationMissingKeyError(
-        `Failed to find entity in in-memory graph object metadata store (_key=${_key})`,
-      );
-    }
-
-    const { _type } = graphObjectMetadata;
     const entities: Entity[] = [];
 
     await this.iterateEntities({ _type }, async (e) => {