+++ packages/integration-sdk-runtime/src/storage/FileSystemGraphObjectStore/__tests__/FileSystemGraphObjectStore.test.ts
@@ -20,7 +20,6 @@ import {
   createIntegrationEntity,
   createIntegrationRelationship,
 } from '@jupiterone/integration-sdk-core';
-import { DuplicateKeyTracker } from '../../../execution';
 
 jest.mock('fs');
 
@@ -258,25 +257,10 @@ describe('getEntity', () => {
       matchingEntity,
     ]);
 
-    const entity = await store.getEntity(_key);
+    const entity = await store.getEntity({ _key, _type });
     expect(store.entityStorageMap.totalItemCount).toEqual(0);
 
     expect(entity).toEqual(matchingEntity);
-  });
-
-  test('should throw if entity is not found by "_key"', async () => {
-    const { storageDirectoryPath, store } = setupFileSystemObjectStore();
-
-    const _type = uuid();
-    const _key = uuid();
-
-    const nonMatchingEntities = times(25, () => generateEntity({ _type }));
-
-    await store.addEntities(storageDirectoryPath, [...nonMatchingEntities]);
-
-    await expect(store.getEntity(_key)).rejects.toThrow(
-      `Failed to find entity in in-memory graph object metadata store (_key=${_key})`,
-    );
   });
 });
 
@@ -410,8 +394,7 @@ describe('iterateRelationships', () => {
 
 function setupFileSystemObjectStore() {
   const storageDirectoryPath = uuid();
-  const duplicateKeyTracker = new DuplicateKeyTracker();
-  const store = new FileSystemGraphObjectStore({ duplicateKeyTracker });
+  const store = new FileSystemGraphObjectStore();
 
   return {
     storageDirectoryPath,